requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
celery-progress==0.1.3
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.7
django-admin-interface==0.20.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-debug-toolbar==3.7.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==3.0.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-select-multiple-field==0.4.2
django-timezone-field==5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
phonenumbers==8.12.54
Pillow==9.2.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
razorpay==1.3.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
